Who can help me?
In a WinForm application, I create a Form with WS_POPUP, WS_EX_NOACTIVE styles, and it has a menu. There is a problem: when I click the menu item, the dropdown is visible. then I click the mouse out the form, the menu's dropdown didn't hide automatically. How can I do?